refactor: decomposed code
