+++ packages/client/src/hooks/useNavigateCustom.ts
