+++ packages/client/src/store/index.ts
@@ -1,1 +1,2 @@
 export { default as store } from './store'
+export type { RootState, AppStore, AppDispatch } from './store'
