+++ packages/client/src/store/reducers/index.ts
@@ -1,0 +1,1 @@
+export { default as authReducer } from './authSlice'
