+++ packages/client/src/store/services/index.ts
@@ -1,0 +1,1 @@
+export { authAPI } from './authService'
